refactor(#528): extract field-management and camera delegates from StorytellingSceneEditor
Extract two delegate classes to reduce StorytellingSceneEditor from 1259 to 1073 lines:

- SceneEditorFieldManager (256 lines): field add/copy/remove statement
  generation, rider detection, vehicle call helpers
- SceneEditorCameraHelper (90 lines): static camera/marker transform
  and color utilities

Override stubs remain on StorytellingSceneEditor to satisfy the
AbstractSceneEditor contract. asSetVehicleCall is static package-private
on SceneEditorFieldManager, shared by both field management and
useSceneAsVehicleForDisconnectedModels.

All 80 characterization tests + full reactor (414 tests, 0 failures) pass.
